requirements.txt
torch
numpy
opencv-python
Pillow
accelerate
git+https://github.com/nullquant/diffusers.git
